# tests/snapshots/web-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Der Diktator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/models": {
      "get": {
        "summary": "Models",
        "operationId": "models_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsStatus"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model}/download": {
      "post": {
        "summary": "Download Model",
        "operationId": "download_model_api_models__model__download_post",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"],
              "type": "string",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsStatus"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/models/{model}/activate": {
      "post": {
        "summary": "Activate Model",
        "operationId": "activate_model_api_models__model__activate_post",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"],
              "type": "string",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsStatus"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/models/{model}/delete": {
      "post": {
        "summary": "Delete Model",
        "operationId": "delete_model_api_models__model__delete_post",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"],
              "type": "string",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsStatus"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/transcribe": {
      "post": {
        "summary": "Transcribe",
        "operationId": "transcribe_api_transcribe_post",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"],
              "type": "string",
              "default": "phonon-2",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transcription"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
...
    "/api/recording-policy": {
      "get": {
        "summary": "Get Recording Policy",
        "operationId": "get_recording_policy_api_recording_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicRecordingPolicy"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["detail", "code"],
        "title": "ApiError",
        "description": "An actionable public failure; context never contains internal diagnostics."
      },
...
      "ClientDeadlines": {
        "properties": {
          "upload": {
            "type": "integer",
            "title": "Upload"
          },
          "batch": {
            "type": "integer",
            "title": "Batch"
          },
          "live": {
            "type": "integer",
            "title": "Live"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["upload", "batch", "live"],
        "title": "ClientDeadlines",
        "description": "Backend-derived maximum client waits in integer milliseconds."
      },
...
      "ModelStatus": {
        "properties": {
          "id": {
            "type": "string",
            "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "languages": {
            "type": "string",
            "title": "Languages"
          },
          "language_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Language Labels"
          },
          "download_mb": {
            "type": "integer",
            "title": "Download Mb"
          },
          "live": {
            "type": "boolean",
            "title": "Live"
          },
          "license": {
            "type": "string",
            "title": "License",
            "default": "CC-BY-4.0"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "installed": {
            "type": "boolean",
            "title": "Installed"
          },
          "state": {
            "type": "string",
            "enum": [
              "missing",
              "installed",
              "downloading",
              "loading",
              "deleting",
              "ready",
              "error"
            ],
            "title": "State"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "languages",
          "language_labels",
          "download_mb",
          "live",
          "source",
          "installed",
          "state"
        ],
        "title": "ModelStatus"
      },
      "ModelsStatus": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "preferred": {
            "type": "string",
            "enum": ["phonon-2", "parakeet-v3", "whisper-large-v3-turbo"],
            "title": "Preferred",
            "default": "phonon-2"
          },
          "busy": {
            "type": "boolean",
            "title": "Busy"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelStatus"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": ["active", "busy", "models"],
        "title": "ModelsStatus"
      },
...
      "PublicRecordingPolicy": {
        "properties": {
          "protocol_version": {
            "type": "integer",
            "const": 1,
            "title": "Protocol Version",
            "default": 1
          },
          "hard_limit_seconds": {
            "type": "integer",
            "title": "Hard Limit Seconds",
            "default": 600
          },
          "wav_container_allowance_bytes": {
            "type": "integer",
            "title": "Wav Container Allowance Bytes",
            "default": 800044
          },
          "max_stream_frame_bytes": {
            "type": "integer",
            "title": "Max Stream Frame Bytes",
            "default": 65536
          },
          "upload_timeout_seconds": {
            "type": "number",
            "title": "Upload Timeout Seconds",
            "default": 180.0
          },
          "batch_timeout_seconds": {
            "type": "number",
            "title": "Batch Timeout Seconds",
            "default": 180.0
          },
          "live_finalization_timeout_seconds": {
            "type": "number",
            "title": "Live Finalization Timeout Seconds",
            "default": 180.0
          },
          "client_timeout_margin_seconds": {
            "type": "number",
            "title": "Client Timeout Margin Seconds",
            "default": 10.0
          },
          "preference_etag": {
            "type": "string",
            "title": "Preference Etag"
          },
          "client_deadlines_ms": {
            "$ref": "#/components/schemas/ClientDeadlines",
            "readOnly": true
          },
          "max_pcm_bytes": {
            "type": "integer",
            "title": "Max Pcm Bytes",
            "readOnly": true
          },
          "max_audio_bytes": {
            "type": "integer",
            "title": "Max Audio Bytes",
            "readOnly": true
          },
          "policy_revision": {
            "type": "string",
            "title": "Policy Revision",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "preference_etag",
          "client_deadlines_ms",
          "max_pcm_bytes",
          "max_audio_bytes",
          "policy_revision"
        ],
        "title": "PublicRecordingPolicy",
        "description": "Agreed limits with the validator of one read of the shared local preferences."
      },
...
      "Transcription": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": ["text"],
        "title": "Transcription",
        "description": "The transcript returned to the browser."
      }
    }
  }
}